Move `Benchmark` to the `benchmark` module
