+++ collector/src/execute/mod.rs
@@ -3,72 +3,26 @@
 use crate::benchmark::patch::Patch;
 use crate::benchmark::profile::Profile;
 use crate::benchmark::scenario::Scenario;
+use crate::benchmark::BenchmarkName;
 use crate::etw_parser;
 use crate::toolchain::Compiler;
 use crate::{command_output, utils};
-use anyhow::{bail, Context};
-use database::{PatchName, QueryLabel};
+use anyhow::Context;
+use database::QueryLabel;
 use futures::stream::FuturesUnordered;
 use futures::stream::StreamExt;
-use std::cmp;
 use std::collections::HashMap;
 use std::env;
-use std::fmt;
-use std::fs::{self, File};
-use std::hash;
+use std::fs;
 use std::io::Read;
-use std::mem::ManuallyDrop;
 use std::path::{Path, PathBuf};
 use std::process::{self, Command};
 use std::str;
 use std::time::Duration;
-use tempfile::TempDir;
 use tokio::runtime::Runtime;
 
 pub mod profiler;
 mod rustc;
-
-fn default_runs() -> usize {
-    3
-}
-
-/// This is the internal representation of an individual benchmark's
-/// perf-config.json file.
-#[derive(Debug, Clone, serde::Deserialize)]
-struct BenchmarkConfig {
-    cargo_opts: Option<String>,
-    cargo_rustc_opts: Option<String>,
-    cargo_toml: Option<String>,
-    #[serde(default)]
-    disabled: bool,
-    #[serde(default = "default_runs")]
-    runs: usize,
-
-    /// The file that should be touched to ensure cargo re-checks the leaf crate
-    /// we're interested in. Likely, something similar to `src/lib.rs`. The
-    /// default if this is not present is to touch all .rs files in the
-    /// directory that `Cargo.toml` is in.
-    #[serde(default)]
-    touch_file: Option<String>,
-
-    category: Category,
-}
-
-#[derive(Ord, PartialOrd, Eq, PartialEq, Clone, Hash)]
-pub struct BenchmarkName(pub String);
-
-impl fmt::Display for BenchmarkName {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-pub struct Benchmark {
-    pub name: BenchmarkName,
-    pub path: PathBuf,
-    pub patches: Vec<Patch>,
-    config: BenchmarkConfig,
-}
 
 // Tools usable with the benchmarking subcommands.
 #[derive(Clone, Copy, Debug, PartialEq)]
@@ -160,27 +114,27 @@ impl PerfTool {
     }
 }
 
-struct CargoProcess<'a> {
-    compiler: Compiler<'a>,
-    cwd: &'a Path,
-    profile: Profile,
-    incremental: bool,
-    processor_etc: Option<(&'a mut dyn Processor, Scenario, &'a str, Option<&'a Patch>)>,
-    processor_name: BenchmarkName,
-    manifest_path: String,
-    cargo_args: Vec<String>,
-    rustc_args: Vec<String>,
-    touch_file: Option<String>,
-    jobserver: Option<jobserver::Client>,
+pub struct CargoProcess<'a> {
+    pub compiler: Compiler<'a>,
+    pub cwd: &'a Path,
+    pub profile: Profile,
+    pub incremental: bool,
+    pub processor_etc: Option<(&'a mut dyn Processor, Scenario, &'a str, Option<&'a Patch>)>,
+    pub processor_name: BenchmarkName,
+    pub manifest_path: String,
+    pub cargo_args: Vec<String>,
+    pub rustc_args: Vec<String>,
+    pub touch_file: Option<String>,
+    pub jobserver: Option<jobserver::Client>,
 }
 
 impl<'a> CargoProcess<'a> {
-    fn incremental(mut self, incremental: bool) -> Self {
+    pub fn incremental(mut self, incremental: bool) -> Self {
         self.incremental = incremental;
         self
     }
 
-    fn processor(
+    pub fn processor(
         mut self,
         processor: &'a mut dyn Processor,
         scenario: Scenario,
@@ -227,7 +181,7 @@ impl<'a> CargoProcess<'a> {
         Ok(package_id.trim().to_string())
     }
 
-    fn jobserver(mut self, server: jobserver::Client) -> Self {
+    pub fn jobserver(mut self, server: jobserver::Client) -> Self {
         self.jobserver = Some(server);
         self
     }
@@ -235,7 +189,7 @@ impl<'a> CargoProcess<'a> {
     // FIXME: the needs_final and processor_etc interactions aren't ideal; we
     // would like to "auto know" when we need final but currently we don't
     // really.
-    fn run_rustc(&mut self, needs_final: bool) -> anyhow::Result<()> {
+    pub fn run_rustc(&mut self, needs_final: bool) -> anyhow::Result<()> {
         log::info!(
             "run_rustc with incremental={}, profile={:?}, scenario={:?}, patch={:?}",
             self.incremental,
@@ -795,249 +749,6 @@ fn store_artifact_sizes_into_stats(stats: &mut Stats, profile: &SelfProfile) {
         stats
             .stats
             .insert(format!("size:{}", artifact.label), artifact.size as f64);
-    }
-}
-
-impl Benchmark {
-    pub fn new(name: String, path: PathBuf) -> anyhow::Result<Self> {
-        let mut patches = vec![];
-        for entry in fs::read_dir(&path)? {
-            let entry = entry?;
-            let path = entry.path();
-            if let Some(ext) = path.extension() {
-                if ext == "patch" {
-                    patches.push(path.clone());
-                }
-            }
-        }
-
-        let mut patches: Vec<_> = patches.into_iter().map(|p| Patch::new(p)).collect();
-        patches.sort_by_key(|p| p.index);
-
-        let config_path = path.join("perf-config.json");
-        let config: BenchmarkConfig = if config_path.exists() {
-            serde_json::from_reader(
-                File::open(&config_path)
-                    .with_context(|| format!("failed to open {:?}", config_path))?,
-            )
-            .with_context(|| format!("failed to parse {:?}", config_path))?
-        } else {
-            bail!("missing a perf-config.json file for `{}`", name);
-        };
-
-        Ok(Benchmark {
-            name: BenchmarkName(name),
-            path,
-            patches,
-            config,
-        })
-    }
-
-    pub fn category(&self) -> Category {
-        self.config.category
-    }
-
-    #[cfg(windows)]
-    fn copy(from: &Path, to: &Path) -> anyhow::Result<()> {
-        crate::utils::fs::robocopy(from, to, &[])
-    }
-
-    #[cfg(unix)]
-    fn copy(from: &Path, to: &Path) -> anyhow::Result<()> {
-        let mut cmd = Command::new("cp");
-        cmd.arg("-pLR").arg(from).arg(to);
-        command_output(&mut cmd)?;
-        Ok(())
-    }
-
-    fn make_temp_dir(&self, base: &Path) -> anyhow::Result<TempDir> {
-        // Appending `.` means we copy just the contents of `base` into
-        // `tmp_dir`, rather than `base` itself.
-        let mut base_dot = base.to_path_buf();
-        base_dot.push(".");
-        let tmp_dir = TempDir::new()?;
-        Self::copy(&base_dot, tmp_dir.path())
-            .with_context(|| format!("copying {} to tmp dir", self.name))?;
-        Ok(tmp_dir)
-    }
-
-    fn mk_cargo_process<'a>(
-        &'a self,
-        compiler: Compiler<'a>,
-        cwd: &'a Path,
-        profile: Profile,
-    ) -> CargoProcess<'a> {
-        let mut cargo_args = self
-            .config
-            .cargo_opts
-            .clone()
-            .unwrap_or_default()
-            .split_whitespace()
-            .map(String::from)
-            .collect::<Vec<_>>();
-        if let Some(count) = env::var("CARGO_THREAD_COUNT")
-            .ok()
-            .and_then(|v| v.parse::<u32>().ok())
-        {
-            cargo_args.push(format!("-j{}", count));
-        }
-
-        CargoProcess {
-            compiler,
-            processor_name: self.name.clone(),
-            cwd,
-            profile,
-            incremental: false,
-            processor_etc: None,
-            manifest_path: self
-                .config
-                .cargo_toml
-                .clone()
-                .unwrap_or_else(|| String::from("Cargo.toml")),
-            cargo_args,
-            rustc_args: self
-                .config
-                .cargo_rustc_opts
-                .clone()
-                .unwrap_or_default()
-                .split_whitespace()
-                .map(String::from)
-                .collect(),
-            touch_file: self.config.touch_file.clone(),
-            jobserver: None,
-        }
-    }
-
-    /// Run a specific benchmark under a processor + profiler combination.
-    pub fn measure(
-        &self,
-        processor: &mut dyn Processor,
-        profiles: &[Profile],
-        scenarios: &[Scenario],
-        compiler: Compiler<'_>,
-        iterations: Option<usize>,
-    ) -> anyhow::Result<()> {
-        let iterations = iterations.unwrap_or(self.config.runs);
-
-        if self.config.disabled || profiles.is_empty() {
-            eprintln!("Skipping {}: disabled", self.name);
-            bail!("disabled benchmark");
-        }
-
-        eprintln!("Preparing {}", self.name);
-        let profile_dirs = profiles
-            .iter()
-            .map(|profile| Ok((*profile, self.make_temp_dir(&self.path)?)))
-            .collect::<anyhow::Result<Vec<_>>>()?;
-
-        // In parallel (but with a limit to the number of CPUs), prepare all
-        // profiles. This is done in parallel vs. sequentially because:
-        //  * We don't record any measurements during this phase, so the
-        //    performance need not be consistent.
-        //  * We want to make use of the reality that rustc is single-threaded
-        //    during a good portion of compilation; that means that it is faster
-        //    to run this preparation when we can interleave rustc's as needed
-        //    rather than fully sequentially, where we have long periods of a
-        //    single CPU core being used.
-        //
-        // As one example, with a full (All profiles x All scenarios)
-        // configuration, script-servo-2 took 2995s without this parallelization
-        // and 2915s with. This is a small win, admittedly, but even a few
-        // minutes shaved off is important -- and there's not too much mangling
-        // of our code needed to get this to work. This benchmark has since been
-        // deleted, but the optimization holds for other crates as well.
-        //
-        // Ideally we would not separately build build-script's (which are
-        // otherwise shared between the configurations), but there's no good way
-        // to do this in Cargo today. We would also ideally build in the same
-        // target directory, but that's also not possible, as Cargo takes a
-        // target-directory global lock during compilation.
-        crossbeam_utils::thread::scope::<_, anyhow::Result<()>>(|s| {
-            let server = jobserver::Client::new(num_cpus::get()).context("jobserver::new")?;
-            for (profile, prep_dir) in &profile_dirs {
-                let server = server.clone();
-                s.spawn::<_, anyhow::Result<()>>(move |_| {
-                    self.mk_cargo_process(compiler, prep_dir.path(), *profile)
-                        .jobserver(server)
-                        .run_rustc(false)?;
-                    Ok(())
-                });
-            }
-            Ok(())
-        })
-        .unwrap()?;
-
-        for (profile, prep_dir) in profile_dirs {
-            eprintln!("Running {}: {:?} + {:?}", self.name, profile, scenarios);
-
-            // We want at least two runs for all benchmarks (since we run
-            // self-profile separately).
-            processor.start_first_collection();
-            for i in 0..cmp::max(iterations, 2) {
-                if i == 1 {
-                    let different = processor.finished_first_collection();
-                    if iterations == 1 && !different {
-                        // Don't run twice if this processor doesn't need it and
-                        // we've only been asked to run once.
-                        break;
-                    }
-                }
-                log::debug!("Benchmark iteration {}/{}", i + 1, iterations);
-                // Don't delete the directory on error.
-                let timing_dir = ManuallyDrop::new(self.make_temp_dir(prep_dir.path())?);
-                let cwd = timing_dir.path();
-
-                // A full non-incremental build.
-                if scenarios.contains(&Scenario::Full) {
-                    self.mk_cargo_process(compiler, cwd, profile)
-                        .processor(processor, Scenario::Full, "Full", None)
-                        .run_rustc(true)?;
-                }
-
-                // Rustdoc does not support incremental compilation
-                if profile != Profile::Doc {
-                    // An incremental  from scratch (slowest incremental case).
-                    // This is required for any subsequent incremental builds.
-                    if scenarios.iter().any(|s| s.is_incr()) {
-                        self.mk_cargo_process(compiler, cwd, profile)
-                            .incremental(true)
-                            .processor(processor, Scenario::IncrFull, "IncrFull", None)
-                            .run_rustc(true)?;
-                    }
-
-                    // An incremental build with no changes (fastest incremental case).
-                    if scenarios.contains(&Scenario::IncrUnchanged) {
-                        self.mk_cargo_process(compiler, cwd, profile)
-                            .incremental(true)
-                            .processor(processor, Scenario::IncrUnchanged, "IncrUnchanged", None)
-                            .run_rustc(true)?;
-                    }
-
-                    if scenarios.contains(&Scenario::IncrPatched) {
-                        for (i, patch) in self.patches.iter().enumerate() {
-                            log::debug!("applying patch {}", patch.name);
-                            patch.apply(cwd).map_err(|s| anyhow::anyhow!("{}", s))?;
-
-                            // An incremental build with some changes (realistic
-                            // incremental case).
-                            let scenario_str = format!("IncrPatched{}", i);
-                            self.mk_cargo_process(compiler, cwd, profile)
-                                .incremental(true)
-                                .processor(
-                                    processor,
-                                    Scenario::IncrPatched,
-                                    &scenario_str,
-                                    Some(&patch),
-                                )
-                                .run_rustc(true)?;
-                        }
-                    }
-                }
-                drop(ManuallyDrop::into_inner(timing_dir));
-            }
-        }
-
-        Ok(())
     }
 }
